Create Shiftable module to move key/offset methods into
